+++ src/pages/apartments/buildings/index.tsx
@@ -1,0 +1,41 @@
+import { useLayoutEffect } from "react";
+import BuildingIcon from "../../../assets/icons/building";
+import { useAppDispatch } from "../../../stores/hooks";
+import { updatePageProperties } from "../../../stores/appFunctionality/pageProperties";
+import BuildingsListsTable from "../../../components/tables/buildings";
+
+const breadCrumb = [
+  {
+    url: "/apartments/view-all",
+    label: "Apartments",
+    icon: <BuildingIcon />,
+  },
+  {
+    url: "#",
+    label: "Buildings",
+    icon: <BuildingIcon />,
+  },
+];
+export default function BuildingsListView() {
+  const dispatch = useAppDispatch();
+  // update page props on component mount
+  useLayoutEffect(() => {
+    dispatch(
+      updatePageProperties({
+        breadCrumb,
+        pageTitle: "Buildings",
+        pageDescription: "Buildings",
+        isLoading: false,
+        failedToLoad: false,
+        setFailedToLoad: false,
+        retryRequest: false,
+      })
+    );
+  }, []);
+
+  return (
+    <div className="w-full">
+      <BuildingsListsTable />
+    </div>
+  );
+}